Alterada propriedade "amount" de int para float (domínio items).
TODO - Corrigir amount de half pizza.

+++ grails-app/domain/capri/Items.groovy
@@ -4,7 +4,7 @@ class Items {
 
 	int id
 	
-	int amount
+	float amount
 	
 	Product product
